% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formstools.R
\docType{data}
\name{form_codebook}
\alias{form_codebook}
\title{An example ODK form schema retrieved using ruODK}
\format{
A tibble with 7 columns and 10 rows:

| **Variable** | **Description** |
| :--- | :--- |
| *path* | XML path to question in ODK |
| *name* | Variable name used in XForm |
| *type* | Variable type |
| *binary* | Are the values for the question binary? |
| *ruodk_name* | Variable name used by ruODK |
| *label* | Question used in XForm |
| *choices* | Named list of choices for select_one and select_multiple questions |
}
\usage{
form_codebook
}
\description{
An example ODK form schema retrieved using ruODK
}
\examples{
form_codebook


}
\keyword{datasets}
