% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formstools.R
\docType{package}
\name{formstools}
\alias{formstools}
\alias{formstools-package}
\title{Tools for working with ODK XLSForms}
\description{
Set of utility functions for use by EcoHealth Alliance researchers in working with Open Data Kit XLSForms. These functions are aimed at aiding users in the data cleaning and data validation process using information found in the ODK XLSForms.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{guevarra@ecohealthalliance.org} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
