% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_choices.R
\name{get_choices}
\alias{get_choices}
\title{Get choices used for select_one and select_multiple questions in an
XLSForm}
\usage{
get_choices(xlsform, choice_name = NULL)
}
\arguments{
\item{xlsform}{A character value for the path to the XLSForm}

\item{choice_name}{A character value or vector of values of names given to the
set of choices in the supplied/specified XLSForm. Default to NULL in which
case all sets of choices are returned.}
}
\value{
A tibble of the `list_name`, `name`, and `label` of the choices
  for the specified `select_one` and `select_multiple` type of questions
  in the specified XLSForm
}
\description{
Get choices used for select_one and select_multiple questions in an
XLSForm
}
\examples{
## Get all choices
get_choices(
  xlsform = system.file(
    "extdata", "ghana_community_form.xlsx", package = "formstools"
  )
)

## Get choices for gender
get_choices(
  xlsform = system.file(
    "extdata", "ghana_community_form.xlsx", package = "formstools"
  ),
  choice_name = "gender"
)


}
