% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruODK_functions.R
\name{match_other_to_choices}
\alias{match_other_to_choices}
\title{Match other responses to specified choices}
\usage{
match_other_to_choices(
  form_schema,
  form_data,
  var_name,
  choice_name = NULL,
  other_var_name,
  id = "id"
)
}
\arguments{
\item{form_schema}{A form schema created by `ruODK::form_scheme_ext()`}

\item{form_data}{A tibble of the data retrieved from ODK Central using ruODK}

\item{var_name}{Specific variable name in ODK form to get choices of}

\item{choice_name}{Specific variable name in `form_schema` that holds
the information on choices. Forms with multiple languages will have a
`form_schema` with separate columns for choices for each language.
Default is NULL which assumes that only one language is used in the form.}

\item{other_var_name}{Variable name in `form_data` for other response in
`var_name`}

\item{id}{A character vector of variable names in `form_data` to use as
identifying data for the output}
}
\value{
A tibble with number of columns the same as the number of possible
  choices for `var_name` in `form_data` plus the number of values in `id`
  used to identify each row of the output. The number of rows is equal to the
  number of rows of data in `form_data`. Names of resulting tibble is the
  concatenation of `recode_` and the choice values for `var_name`
}
\description{
Match other responses to specified choices
}
\examples{
## Match other toppings responses (pizza3) to choices for
## pizza toppings (pizza2)
match_other_to_choices(
  form_schema = form_codebook, form_data = pizza_data,
  var_name = "pizza2", other_var_name = "pizza3"
)


}
