% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formstools.R
\docType{data}
\name{pizza_data}
\alias{pizza_data}
\title{An example data collected using ODK and aggregated using ODK Central}
\format{
A tibble with 21 columns and 17 rows:

| **Variable** | **Description** |
| :--- | :--- |
| *id* | Identifier |
| *start* | Start date and time |
| *end* | End date and time |
| *today* | Date of data entry |
| *pizza1* | Do you like pizza? |
| *pizza2* | Pizza toppings you like (select multiple) |
| *pizza3* | Any other toppings? |
| *closing* | Closing statement |
| *meta_audit* | Meta audit |
| *meta_instance_id* | Meta instance ID |
| *system_submission_date* | System submission date |
| *system_updated_at* | System updated at |
| *system_submitter_id* | System submitter ID |
| *system_submitter_name* | System submitter name |
| *system_attachments_present* | System attachments present |
| *system_attachments_expected* | System attachments expected |
| *system_status* | System status |
| *system_review_state* | System review state |
| *system_device_id* | System device ID |
| *system_edits* | System edits |
| *odata_context* | OData context |
}
\usage{
pizza_data
}
\description{
An example data collected using ODK and aggregated using ODK Central
}
\examples{
pizza_data

}
\keyword{datasets}
