% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spread_vector_to_columns}
\alias{spread_vector_to_columns}
\title{Convert character vector of categorical responses into unique variables}
\usage{
spread_vector_to_columns(x, fill = NULL, na_rm = FALSE, prefix)
}
\arguments{
\item{x}{Vector of categorical values}

\item{fill}{A vector of all the possible responses to the select multiple
question}

\item{na_rm}{Logical. Should an NA response be reported in its own column?
Default to FALSE.}

\item{prefix}{A character string to prepend to the names of the new columns
to be created}
}
\value{
A tibble with number of columns the same as the number of unique
  values `x` and number of rows equal to the length of `x`. Names of
  resulting tibble is the concatenation of the `prefix` and the values of `x`
}
\description{
Function transforms a vector of categorical responses into `n` number of
new columns/variables equal to the number of unique categorical values.
}
\examples{
spread_vector_to_columns(
  x = c("cat", "cat", "dog", "dog", "dog", NA_character_),
  fill = c("cat", "dog", "monkey"),
  na_rm = TRUE,
  prefix = "pets"
)

}
