% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruODK_functions.R
\name{get_choices_ruodk}
\alias{get_choices_ruodk}
\title{Get choices for a specified variable in an ODK form}
\usage{
get_choices_ruodk(form_schema, var_name = NULL)
}
\arguments{
\item{form_schema}{A form schema created by `ruODK::form_scheme_ext()`}

\item{var_name}{Specific variable name to get choices of}
}
\value{
A tibble with `var_name`, `question`, `values`, and `labels`
}
\description{
Get choices for a specified variable in an ODK form
}
\examples{
## Get all choices for select_one and select_multiple type questions
get_choices_ruodk(form_schema = form_codebook)

## Get choices for pizza2 variable
get_choices_ruodk(form_schema = form_codebook, var_name = "pizza2")

}
