% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_multiple_response}
\alias{split_multiple_response}
\alias{split_multiple_responses}
\title{Split a vector of values from an ODK select multiple type of response}
\usage{
split_multiple_response(x, sep = " ", fill, na_rm = FALSE, prefix)

split_multiple_responses(x, sep = " ", fill, na_rm = FALSE, prefix)
}
\arguments{
\item{x}{A select_multiple response or vector of responses with multiple
responses}

\item{sep}{Separator used to separate multiple responses. Default to " ".
Regular expressions can be used to detect more than one possible separator.}

\item{fill}{A vector of all the possible responses to the select multiple
question}

\item{na_rm}{Logical. Should an NA response be reported in its own column?
Default to FALSE.}

\item{prefix}{A character value for prefix to append to names of resulting
data.frame.}
}
\value{
A tibble with number of columns the same as the number of possible
  choices for the select multiple question (same as lenght of `fill`) and
  number of rows equal to the length of `x`. Names of resulting tibble is the
  concatenation of the `prefix` and the values for `fill`
}
\description{
Split a vector of values from an ODK select multiple type of response
}
\examples{
## Split the multiple responses of pizza toppings
split_multiple_responses(
  x = pizza_data$pizza2,
  sep = ", ",
  fill = c("cheese", "tomatoes", "pepperoni", "mushrooms",
           "artichoke", "olives", "pineapple", "other"),
  prefix = "toppings"
)

}
